Remove usages of elasticsearch.build plugin in non-production projects
