+++ x-pack/plugin/eql/qa/correctness/build.gradle
@@ -1,7 +1,6 @@
+apply plugin: 'elasticsearch.java'
 apply plugin: 'elasticsearch.internal-java-rest-test'
-apply plugin: 'elasticsearch.build'
 apply plugin: 'elasticsearch.internal-testclusters'
-tasks.named("test").configure { enabled = false }
 
 import org.elasticsearch.gradle.testclusters.RunTask
 import org.elasticsearch.gradle.internal.info.BuildParams
@@ -23,11 +22,6 @@ Boolean preserveData = providers.systemProperty('eql.test.preserve.data')
   .map { s -> Boolean.parseBoolean(s) }
   .getOrElse(false)
 
-if (BuildParams.inFipsJvm){
-  // This test cluster is using a BASIC license and FIPS 140 mode is not supported in BASIC
-  tasks.named("javaRestTest").configure{ enabled = false }
-}
-
 testClusters.configureEach {
     if (serviceAccountFile) {
       keystore 'gcs.client.eql_test.credentials_file', serviceAccountFile
@@ -47,7 +41,7 @@ def runTaskCluster = testClusters.register('runTask') {
 }
 
 tasks.named('javaRestTest').configure {
-  onlyIf { serviceAccountFile }
+  onlyIf { serviceAccountFile && BuildParams.inFipsJvm == false }
 
   testLogging {
     showStandardStreams = true
