+++ x-pack/plugin/security/qa/operator-privileges-tests/build.gradle
@@ -1,6 +1,6 @@
 import org.elasticsearch.gradle.Version
 
-apply plugin: 'elasticsearch.internal-es-plugin'
+apply plugin: 'elasticsearch.base-internal-es-plugin'
 apply plugin: 'elasticsearch.internal-java-rest-test'
 
 esplugin {
