+++ x-pack/qa/freeze-plugin/build.gradle
@@ -8,7 +8,6 @@
 
 apply plugin: 'elasticsearch.internal-es-plugin'
 
-
 esplugin {
   name = 'freeze-plugin'
   description = 'Provides freeze-index endpoint for testing purposes only'
